fix: Complete type safety improvements for core library
Major type safety overhaul achieving 100% type safety in core library modules:

Fixed Type Errors:
- robust_inference.py: Fixed list-to-array reassignments using new variables
- stability.py: Fixed list-to-array reassignments
- weights.py: Fixed Hill estimator return type and list annotations
- overlap.py: Explicit float conversions for tuple returns
- dr_dashboards.py: Added None check for ifs_data
- weight_dashboards.py: Added None checks and fixed tuple type for tight_layout
- fresh_draw_utils.py: Added hasattr checks for metadata/reward attributes
- dataset.py: Used cast() for Literal type constraints
- iic.py: Explicit float conversion for numpy mean
- sampling.py: Added np.asarray() wrappers for proper return types
- oracle_slice.py: Wrapped isotonic prediction in np.asarray()
- precomputed_sampler.py: Used cast() for PolicyDataDict type

Type Safety Improvements:
- Replaced reassignment patterns (list = np.array(list)) with new variables
- Added explicit type annotations for List variables
- Used cast() instead of type: ignore for better documentation
- Added proper None/Optional handling throughout
- Explicit float() conversions for numpy scalars

Results:
- Reduced total mypy errors from 269 to 183 (32% reduction)
- Core library (non-test/experiment) errors: 0 (100% type safe!)
- Remaining errors are in test files and experiments only

The core CJE library is now fully type-safe, providing better IDE support,
catching more bugs at development time, and improving code maintainability.

+++ cje/calibration/iic.py
@@ -260,7 +260,9 @@ class IsotonicInfluenceControl:
 
             direction_counts = Counter(directions_used)
             self._last_direction = direction_counts.most_common(1)[0][0]
-            self._last_correlation = np.mean(correlations) if correlations else 0.0
+            self._last_correlation = (
+                float(np.mean(correlations)) if correlations else 0.0
+            )
             self._fold_directions = directions_used  # Store for detailed diagnostics
 
         return fitted
